Split out Page Container into its own file
