+++ components/brave_wallet_ui/page/wallet_page.tsx
@@ -5,10 +5,10 @@
 
 import * as React from 'react'
 import { render } from 'react-dom'
-import { connect, Provider } from 'react-redux'
-import { bindActionCreators, Dispatch } from 'redux'
+import { Provider } from 'react-redux'
 import { initLocale } from 'brave-ui'
 
+import Container from './container'
 import * as WalletPageActions from './actions/wallet_page_actions'
 import store from './store'
 
@@ -16,85 +16,12 @@ import 'emptykit.css'
 import '../../../ui/webui/resources/fonts/poppins.css'
 import '../../../ui/webui/resources/fonts/muli.css'
 
-import { WalletWidgetStandIn } from '../stories/style'
-import {
-  SideNav,
-  WalletPageLayout,
-  WalletSubViewLayout,
-  CryptoView
-} from '../components/desktop'
 import LegacyApp from '../components/legacy_app'
-import {
-  NavTypes,
-  NavObjectType,
-  WalletState,
-  PageState,
-  WalletPageState
-} from '../constants/types'
-import { LinkedAccountsOptions, NavOptions, StaticOptions } from '../options/side-nav-options'
-import BuySendSwap from '../components/buy-send-swap'
 import Theme from 'brave-ui/theme/brave-default'
 import DarkTheme from 'brave-ui/theme/brave-dark'
 import BraveCoreThemeProvider from '../../common/BraveCoreThemeProvider'
 import walletDarkTheme from '../theme/wallet-dark'
 import walletLightTheme from '../theme/wallet-light'
-
-type Props = {
-  wallet: WalletState
-  page: PageState
-  actions: typeof WalletPageActions
-}
-
-function Page (props: Props) {
-  const [view, setView] = React.useState<NavTypes>('crypto')
-  const [linkedAccounts] = React.useState<NavObjectType[]>(LinkedAccountsOptions)
-
-  // In the future these will be actual paths
-  // for example wallet/rewards
-  const navigateTo = (path: NavTypes) => {
-    setView(path)
-  }
-
-  return (
-    <WalletPageLayout>
-        <SideNav
-          navList={NavOptions}
-          staticList={StaticOptions}
-          selectedButton={view}
-          onSubmit={navigateTo}
-          linkedAccountsList={linkedAccounts}
-        />
-        <WalletSubViewLayout>
-          {view === 'crypto' ? (
-            <CryptoView />
-          ) : (
-            <div style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
-              <h2>{view} view</h2>
-            </div>
-          )}
-
-        </WalletSubViewLayout>
-        <WalletWidgetStandIn>
-          <BuySendSwap />
-        </WalletWidgetStandIn>
-      </WalletPageLayout>
-  )
-}
-
-function mapStateToProps (state: WalletPageState): Partial<Props> {
-  return {
-    page: state.page,
-    wallet: state.wallet
-  }
-}
-
-function mapDispatchToProps (dispatch: Dispatch): Partial<Props> {
-  return {
-    actions: bindActionCreators(WalletPageActions, store.dispatch.bind(store))
-  }
-}
-
-const PageWithState = connect(mapStateToProps, mapDispatchToProps)(Page)
 
 function App () {
   const [initialThemeType, setInitialThemeType] = React.useState<chrome.braveTheme.ThemeType>()
@@ -109,8 +36,7 @@ function App () {
         dark={walletDarkTheme}
         light={walletLightTheme}
       >
-        <PageWithState
-        />
+        <Container />
       </BraveCoreThemeProvider>
       }
     </Provider>
